Add update display on keypad button clicks
Right now, clicking a keypad button simply
appends the clicked button's value to the main
display string regardless of the string content.
The function handling this will need to include
logic to account for various situations in which
the main display string should behave differently
to certain keypad button clicks.